fix(tests): extract shared TempDirectory and StubHttpMessageHandler helpers
QLTY flagged 4 test fixture duplication smells. Extract shared helpers:
- Common/TempDirectory.cs — used by SupportBundle, SpawnPreset, ModOnboarding tests
- Common/StubHttpMessageHandler.cs — used by ProfileUpdate tests

Also adds .coverage-thresholds.json (100% lines/branches/functions/statements)
required by metaswarm coverage gate.
