+++ .coverage-thresholds.json
@@ -1,0 +1,7 @@
+{
+  "lines": 100,
+  "branches": 100,
+  "functions": 100,
+  "statements": 100,
+  "command": "pwsh tools/quality/assert-dotnet-coverage.ps1 -CoveragePath TestResults/coverage/cobertura.xml"
+}
