+++ tests/SwfocTrainer.Tests/Core/SupportBundleCoverageSweepTests.cs
@@ -1,4 +1,5 @@
 using System.IO.Compression;
+using SwfocTrainer.Tests.Common;
 using System.Text.Json;
 using FluentAssertions;
 using SwfocTrainer.Core.Contracts;
@@ -96,24 +97,6 @@ public sealed class SupportBundleCoverageSweepTests
         }
     }
 
-    private sealed class TempDirectory : IDisposable
-    {
-        public TempDirectory(string prefix)
-        {
-            Path = System.IO.Path.Combine(System.IO.Path.GetTempPath(), $"{prefix}-{Guid.NewGuid():N}");
-            Directory.CreateDirectory(Path);
-        }
-
-        public string Path { get; }
-
-        public void Dispose()
-        {
-            if (Directory.Exists(Path))
-            {
-                Directory.Delete(Path, recursive: true);
-            }
-        }
-    }
 
     private sealed class RunFixture : IDisposable
     {
