+++ tests/SwfocTrainer.Tests/Profiles/ProfileUpdateServiceCoverageSweepTests.cs
@@ -3,6 +3,7 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Text.Json;
+using SwfocTrainer.Tests.Common;
 using FluentAssertions;
 using SwfocTrainer.Core.Contracts;
 using SwfocTrainer.Core.Models;
@@ -191,36 +192,6 @@ public sealed class ProfileUpdateServiceCoverageSweepTests
 
     private static string ComputeSha256(byte[] bytes)
         => Convert.ToHexString(System.Security.Cryptography.SHA256.HashData(bytes)).ToLowerInvariant();
-
-    private sealed class StubHttpMessageHandler : HttpMessageHandler
-    {
-        private readonly IReadOnlyDictionary<string, (string ContentType, byte[] Body)> _responses;
-
-        public StubHttpMessageHandler(IReadOnlyDictionary<string, (string ContentType, byte[] Body)> responses)
-        {
-            _responses = responses;
-        }
-
-        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-        {
-            _ = cancellationToken;
-            var key = request.RequestUri!.ToString();
-            if (!_responses.TryGetValue(key, out var payload))
-            {
-                return Task.FromResult(new HttpResponseMessage(HttpStatusCode.NotFound)
-                {
-                    Content = new StringContent("not found")
-                });
-            }
-
-            var response = new HttpResponseMessage(HttpStatusCode.OK)
-            {
-                Content = new ByteArrayContent(payload.Body)
-            };
-            response.Content.Headers.ContentType = new System.Net.Http.Headers.MediaTypeHeaderValue(payload.ContentType);
-            return Task.FromResult(response);
-        }
-    }
 
     private sealed class ThrowingHttpMessageHandler : HttpMessageHandler
     {
